requirements.txt
Django==1.8.2
flake8==2.4.1
ipdb==0.8.1
ipython==3.2.0
mccabe==0.3.1
pbr==1.5.0
pep8==1.5.7
psycopg2==2.6.1
pyflakes==0.8.1
PyYAML==3.11
requests==2.7.0
six==1.9.0
vcrpy==1.7.2
wheel==0.24.0
wrapt==1.10.5
gunicorn==19.3.0
dj-database-url==0.3.0
whitenoise==2.0.3
